Construct AAGI Colour Scales for ggplot2

Uses AAGI colours to create colour scales suitable for use in ggplot2 objects.

Usage

scale_colour_aagi(palette = "colourful", discrete = TRUE, reverse = FALSE, ...)

Arguments

palette

Character string, one of:

"primary",

"secondary",

"colourful" (default),

"full",

"bright", or

"bluegreen".

discrete

Boolean argument indicating whether colour aesthetic is discrete, e.g., a factor (TRUE) or continuous, e.g., numeric (FALSE).

reverse

Boolean argument indicating whether the palette should be reversed, i.e., light to dark. Defaults to FALSE.

...

Additional arguments passed to discrete_scale() or ggplot2::scale_colour_gradientn, used respectively when discrete is TRUE.

See also